Accelerate corpus evolution by merge control files
In libfuzzer, it generate the covered features and edges for each corpora and saves them in the merge_control_file.

The merge process of libfuzzer is underwent by adding corpora with new features added.

See in https://reviews.llvm.org/D66107.

In this commit, relying on the merge control file, we store the set of features that covered by the shared corpus.
When meet new corpora, we examize whether this corpora brings new features.
If new feature gained, we add this corpora into the shared corpus.

By this way, we can minimize the CPU cost by avoid redundant and heavily coverage computions.