Constrain the version of sqlite3 in its build script
In the recent version of sqlite3, the build command "make sqlite3.c" does not produce the "sqlite3.o" that PromptFuzz required anymore.

So I fix the version of sqlite3 to an early version in its build script.
If someone need to use the latest version of sqlite3, please modify the build script and generate the "sqlite3.o".